test: Move Dummy to testCore module
