+++ main/test/src/io/github/iltotore/iron/test/main/CompileTimeSpec.scala
@@ -1,7 +1,9 @@
-package io.github.iltotore.iron.test
+package io.github.iltotore.iron.test.main
 
 import org.scalatest._, flatspec._, matchers._
 import io.github.iltotore.iron._, constraint.{_, given}
+
+import io.github.iltotore.iron.test.{Dummy, UnitSpec, given}
 
 class CompileTimeSpec extends UnitSpec {
