+++ main/test/src/io/github/iltotore/iron/test/main/package.scala
@@ -1,14 +1,8 @@
-package io.github.iltotore.iron
+package io.github.iltotore.iron.test
 
 import io.github.iltotore.iron.constraint.Constraint
 
-package object test {
-
-  trait Dummy
-
-  inline given Constraint[Boolean, Dummy] with {
-    override inline def assert(value: Boolean): Boolean = value
-  }
+package object main {
 
   trait DummyCompileTime
